+++ frontend/.eslintrc.json
@@ -56,7 +56,8 @@
         "react/jsx-no-useless-fragment": "off",
         "@typescript-eslint/no-unused-vars": "error",
         "no-unused-vars": "off",
-        "no-shadow": "off"
+        "no-shadow": "off",
+        "no-restricted-exports": "off"
     },
     "settings": {
         "import/resolver": {
@@ -68,7 +69,7 @@
                 "map": [
                     ["@components", "./src/components"],
                     ["@contexts", "./src/contexts"],
-                    ["@lib", "./src/lib"],
+                    ["@utils", "./src/utils"],
                     ["@models", "./src/models"],
                     ["@test", "./src/test"]
                 ]
