+++ frontend/src/components/Pages/NoAccess/NoAccess.tsx
@@ -1,0 +1,7 @@
+import { Alert } from 'antd';
+
+const NoAccess = (): JSX.Element => (
+    <Alert type="error" message="No Access" />
+);
+
+export default NoAccess;
