--- frontend/src/lib/copyObject.ts
@@ -1,3 +1,0 @@
-export function copyObject<T>(obj: T): T {
-    return JSON.parse(JSON.stringify(obj));
-}
