+++ frontend/src/utils/copyObject.ts
@@ -1,0 +1,3 @@
+export default function copyObject<T>(obj: T): T {
+    return JSON.parse(JSON.stringify(obj));
+}
